requirements.txt
discord.py
python-dotenv
aiohttp
matplotlib
numpy
pandas
scikit-learn
statsmodels
prophet
edge-tts
wikipedia
opencv-python
fastapi
uvicorn
plotly
aiosqlite
pytz
python-whois
watchdog
cryptography
transformers
torch
protobuf
fugashi
unidic-lite
sentencepiece
prometheus_client
jwt